+++ packages/adminui-vue/utils/Form/validation.js
@@ -3,6 +3,7 @@ module.exports = createValidator
 const Vue = require('vue')
 const { validationMixin } = require('vuelidate/lib/index')
 const { required } = require('vuelidate/lib/validators/index')
+const { mapInstanceFields } = require('./helpers')
 
 /**
  * Track Instance module data and validate it
@@ -10,35 +11,33 @@ const { required } = require('vuelidate/lib/validators/index')
  * props which have !allowNull param and defaultView is true
  * @param {Vuex} store Store
  * @param {UBEntity} entitySchema Entity schema
- * @param {Array<string>} fieldList Master request field list
+ * @param {Vue} customValidationMixin Custom validation mixin in case when need to override default validation
  * @return {object} Vuelidate validation object
  */
-function createValidator (store, entitySchema, fieldList) {
+function createValidator (store, entitySchema, customValidationMixin = {}) {
   const requiredFields = entitySchema
     .filterAttribute(attr => attr.defaultView && !attr.allowNull)
     .map(a => a.name)
 
-  const validatorInstance = new Vue({
-    store,
-    mixins: [ validationMixin ],
-
-    computed: {
-      values () {
-        return this.$store.state.data
-      }
-    },
+  const defaultValidationMixin = {
+    computed: mapInstanceFields(requiredFields),
 
     validations () {
-      return {
-        values: fieldList.reduce((obj, field) => {
-          const isRequired = requiredFields.includes(field)
-          obj[field] = isRequired ? { required } : {}
-
-          return obj
-        }, {})
-      }
+      return requiredFields.reduce((obj, field) => {
+        obj[field] = { required }
+        return obj
+      }, {})
     }
+  }
+
+  const instance = new Vue({
+    store,
+    mixins: [
+      validationMixin,
+      defaultValidationMixin,
+      customValidationMixin
+    ]
   })
 
-  return validatorInstance.$v.values
+  return instance.$v
 }